+++ admin/projects/backend_projects.php
@@ -9,6 +9,16 @@
 
         if (isset($_GET['delete'])) {
             $delete_id = intval($_GET['delete']);
+
+            $sql = "SELECT * FROM projects WHERE id = $delete_id"; //Note: select all // 
+            $result = mysqli_query($conn,$sql);
+            
+            if (mysqli_num_rows($result) == 1) {
+                $get = mysqli_fetch_assoc($result); // Note: getting the result into an array //
+                $file_name = $get['file_name']; 
+                unlink($_SERVER['DOCUMENT_ROOT'] . "/my_portfolio/uploads/projects/" . $user_id . "/".DIRECTORY_SEPARATOR. $file_name);
+            }
+
             $sql = "DELETE FROM projects WHERE id = $delete_id"; // important for delete // 
             if ($conn->query($sql) === TRUE) {
                 $_SESSION['deleted_message'] = "Successfully Deleted";
@@ -29,20 +39,18 @@
             $status = test_input($_POST["status"]);
             $file_name = basename($_FILES["fileToUpload"]["name"]);
 
-            if(isset($_POST["submit"]) && !empty(basename($_FILES["fileToUpload"]["name"]))) {
+            if(!empty(basename($_FILES["fileToUpload"]["name"]))) {
                 $target_dir = $_SERVER['DOCUMENT_ROOT']. "/my_portfolio/uploads/projects/" . $user_id . "/";
                 $target_file = $target_dir . basename($_FILES["fileToUpload"]["name"]);
                 $uploadOk = 1;
                 $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
                 $check = getimagesize($_FILES["fileToUpload"]["tmp_name"]);
-                
-                // header("Location: ../about/about.php");
-                // exit;
+
                 if($check !== false) {
                     $uploadOk = 1; 
                 } else {
                     $uploadOk = 0;
-                    $_SESSION['error'] = 'Image Error, Please Upload Again';
+                    $_SESSION['error'] = 'Image must be PNG, JPEG, JPG, or any JFIF files, Please Upload Again';
                     header("Location: ../projects/projects.php");
                     exit;
                 }
@@ -51,28 +59,30 @@
                     mkdir($target_dir, 0777, true);
                 }
 
+                if (file_exists($target_file)) {
+                    $uploadOk = 0;
+                    $_SESSION['error'] = "Sorry, file already exists.";
+                    header("Location: ../projects/projects.php");
+                    exit;
+                    
+                }
+
                 if ($uploadOk == 0) {
                 } else {
                     if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)) {
-                        echo 'try';
-                        exit;
                     } else {
                         echo 'no';
                         exit;
                     }
                 }
-            }
-       
 
-        if ($file_name) {
-           $sql = "INSERT INTO projects (name, description, status, user_id, file_name) VALUES ('$project_name', '$description', '$status', '$user_id', '$file_name')";
-           $_SESSION['success_message'] = "Successfully Created a Project!";
-        } else {
-            $sql = "INSERT INTO projects (name, description, status, user_id) VALUES ('$project_name', '$description', '$status', '$user_id')";
-            $_SESSION['success_message'] = "Successfully Created a Project!";
-        }
-        // $sql = "INSERT INTO projects (name, description, status, user_id, file_name) VALUES ('$project_name', '$description', '$status', '$user_id', '$file_name')"; // important for create // 
-        
+                $sql = "INSERT INTO projects (name, description, status, user_id, file_name) VALUES ('$project_name', '$description', '$status', '$user_id', '$file_name')";
+                $_SESSION['success_message'] = "Successfully Created a Project!";
+            } else {
+                $sql = "INSERT INTO projects (name, description, status, user_id) VALUES ('$project_name', '$description', '$status', '$user_id')";
+                $_SESSION['success_message'] = "Successfully Created a Project!";
+            }
+                    
         if ($conn->query($sql) === TRUE) {
             //$_SESSION['success_message'] = "Successfully Created a Project!";
         } else {
